requirements.txt
# requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pyngrok==7.1.6
elasticsearch==8.18.0
nest_asyncio==1.6.0
